+++ src/main/java/com/ddoddo/balllog/balllog/dto/request/BallLogRequest.java
@@ -3,8 +3,8 @@ package com.ddoddo.balllog.balllog.dto.request;
 import java.time.LocalDateTime;
 
 public interface BallLogRequest {
-    Long cheeringTeamId();
-    Long opposingTeamId();
+    Integer cheeringTeamId();
+    Integer opposingTeamId();
     Integer scoreCheering();
     Integer scoreOpposing();
     String title();
