+++ src/main/java/com/ddoddo/balllog/kbo/repository/KboTeamRepository.java
@@ -3,5 +3,5 @@ package com.ddoddo.balllog.kbo.repository;
 import com.ddoddo.balllog.kbo.model.KboTeam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KboTeamRepository extends JpaRepository<KboTeam, Long> {
+public interface KboTeamRepository extends JpaRepository<KboTeam, Integer> {
 }
